requirements.txt
django
djangorestframework
flask
s3transfer
Werkzeug
whitenoise
zappa-django-utils